Implemented AuthorizeExceptSuspended attribute; however, it has a a design problem. If you do [AuthorizeExceptSuspended(Roles="Admin")], it checks for admin and doesn't check whether the user is suspended before it.

+++ FindPianos/FindPianos/Models/Validation.cs
@@ -15,7 +15,7 @@ namespace FindPianos.Models
 
         public IEnumerable<RuleViolation> GetRuleViolations()
         {
-            if ((int)RatingOverall == null )
+            if (RatingOverall == null)
                 yield return new RuleViolation("Overall rating is required", "RatingOverall");
 
             yield break;
